Moving Camera//TODO::fix player movement

+++ GameProduction 2/GameProduction 2/Player.cpp
@@ -24,11 +24,11 @@ void Player::Update()
 	m_velX += m_accelX;
 	m_velX *= (m_grounded ? m_drag : 1);
 	m_velX = std::min(std::max(m_velX, -(m_maxVelX)), (m_maxVelX));
-	m_dst.x += (int)m_velX; // Had to cast it to int to get crisp collision with side of platform.
+	m_dst.x += m_velX; // Had to cast it to int to get crisp collision with side of platform.
 	// Now do Y axis.
 	m_velY += m_accelY + m_grav; // Adjust gravity to get slower jump.
 	m_velY = std::min(std::max(m_velY, -(m_maxVelY)), (m_grav * 3));
-	m_dst.y += (int)m_velY; // If you run into issues with vertical collision, you can also cast to int.
+	m_dst.y += m_velY; // If you run into issues with vertical collision, you can also cast to int.
 	m_accelX = m_accelY = 0.0;
 	if (EVMA::KeyHeld(SDL_SCANCODE_A))
 		m_accelX -= 1.0;
